chore: Test and build only updated packages

+++ .github/actions/rebuild-docs/entrypoint.sh
@@ -25,9 +25,14 @@ REPO="${REPO/https:\/\/github.com\//https:\/\/${GH_USER}:${GH_TOKEN}@github.com\
 git remote set-url origin "${REPO}"
 
 rm -rf docs/packages/*
-yarn build:docs
+./bin/updated.sh "build:docs"
+
+if [ "$?" != "0" ]; then
+  exit $?
+fi
+
 git add docs
-git commit -m "docs: Rebuild docs"
+git commit -m "docs: Rebuild docs [ci skip]"
 git push origin master
 
 rm "${HOME}/.gitconfig"
